fix code warning

+++ frameworks/core/components_ng/pattern/calendar/calendar_paint_method.cpp
@@ -707,7 +707,7 @@ bool CalendarPaintMethod::IsOffDay(const CalendarDay& dayInfo) const
     StringUtils::StringSplitter(offDays_, ',', days);
     int32_t daysOfWeek = 7;
     for (const auto& day : days) {
-        auto num = (StringUtils::StringToInt(day) + (daysOfWeek - startOfWeek_)) % daysOfWeek;
+        auto num = (StringUtils::StringToInt(day) + (daysOfWeek - static_cast<int32_t>(startOfWeek_))) % daysOfWeek;
         if ((dayInfo.index % daysOfWeek) == num) {
             return true;
         }
